[dice] move keygen function to cert lib
This moves the ECC P256 keygen wrapper function to the cert lib, since
asymmetric keys destined for wrapping via certs can be either DICE, TPM,
or other yet-to-be categorized keys.
