+++ sw/device/silicon_creator/lib/base/util.c
@@ -12,3 +12,12 @@ uint32_t util_round_up_to(uint32_t input, uint32_t align_bits) {
 uint32_t util_size_to_words(uint32_t bytes) {
   return util_round_up_to(bytes, 2) / sizeof(uint32_t);
 }
+
+void util_le_be_buf_format(unsigned char *buf, size_t num_bytes) {
+  unsigned char temp;
+  for (size_t i = 0; i < (num_bytes / 2); ++i) {
+    temp = buf[i];
+    buf[i] = buf[num_bytes - i - 1];
+    buf[num_bytes - i - 1] = temp;
+  }
+}
